+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/datastructures/state/BeaconStateImpl.java
@@ -11,10 +11,11 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.spec.containers.state;
+package tech.pegasys.teku.spec.datastructures.state;
 
 import java.util.Optional;
 import java.util.function.Consumer;
+import java.util.function.Function;
 import org.apache.tuweni.bytes.Bytes32;
 import tech.pegasys.teku.datastructures.blocks.BeaconBlockHeader;
 import tech.pegasys.teku.datastructures.blocks.Eth1Data;
@@ -25,140 +26,166 @@ import tech.pegasys.teku.datastructures.state.PendingAttestation;
 import tech.pegasys.teku.datastructures.state.Validator;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 import tech.pegasys.teku.ssz.SSZTypes.Bitvector;
+import tech.pegasys.teku.ssz.SSZTypes.SSZBackingList;
 import tech.pegasys.teku.ssz.SSZTypes.SSZList;
 import tech.pegasys.teku.ssz.SSZTypes.SSZVector;
+import tech.pegasys.teku.ssz.backing.SszContainer;
+import tech.pegasys.teku.ssz.backing.SszData;
+import tech.pegasys.teku.ssz.backing.cache.IntCache;
+import tech.pegasys.teku.ssz.backing.schema.SszCompositeSchema;
+import tech.pegasys.teku.ssz.backing.schema.SszContainerSchema;
+import tech.pegasys.teku.ssz.backing.tree.TreeNode;
+import tech.pegasys.teku.ssz.backing.view.SszContainerImpl;
 import tech.pegasys.teku.ssz.backing.view.SszPrimitives;
 
-public class DelegatingBeaconState implements BeaconState {
-
-  protected final BeaconState state;
-
-  public DelegatingBeaconState(final BeaconState state) {
-    this.state = state;
+class BeaconStateImpl extends SszContainerImpl implements BeaconState {
+  public BeaconStateImpl(BeaconStateSchema schema) {
+    super(schema);
   }
+
+  BeaconStateImpl(SszCompositeSchema<?> type, TreeNode backingNode, IntCache<SszData> cache) {
+    super(type, backingNode, cache);
+  }
+
+  BeaconStateImpl(SszContainerSchema<? extends SszContainer> type, TreeNode backingNode) {
+    super(type, backingNode);
+  }
+
+  // Example getters
 
   @Override
   public UInt64 getGenesis_time() {
-    return state.getGenesis_time();
-  }
-
-  @Override
-  public Bytes32 getGenesis_validators_root() {
-    return state.getGenesis_validators_root();
-  }
-
-  @Override
-  public UInt64 getSlot() {
-    return state.getSlot();
-  }
-
-  @Override
-  public Fork getFork() {
-    return state.getFork();
-  }
-
-  @Override
-  public ForkInfo getForkInfo() {
-    return state.getForkInfo();
-  }
-
-  @Override
-  public BeaconBlockHeader getLatest_block_header() {
-    return state.getLatest_block_header();
-  }
-
-  @Override
-  public SSZVector<Bytes32> getBlock_roots() {
-    return state.getBlock_roots();
-  }
-
-  @Override
-  public SSZVector<Bytes32> getState_roots() {
-    return state.getState_roots();
-  }
-
-  @Override
-  public SSZList<Bytes32> getHistorical_roots() {
-    return state.getHistorical_roots();
-  }
-
-  @Override
-  public Eth1Data getEth1_data() {
-    return state.getEth1_data();
-  }
-
-  @Override
-  public SSZList<Eth1Data> getEth1_data_votes() {
-    return state.getEth1_data_votes();
-  }
-
-  @Override
-  public UInt64 getEth1_deposit_index() {
-    return state.getEth1_deposit_index();
-  }
-
-  @Override
-  public SSZList<Validator> getValidators() {
-    return state.getValidators();
-  }
-
-  @Override
-  public SSZList<UInt64> getBalances() {
-    return state.getBalances();
-  }
-
-  @Override
-  public SSZVector<Bytes32> getRandao_mixes() {
-    return state.getRandao_mixes();
-  }
-
-  @Override
-  public SSZVector<UInt64> getSlashings() {
-    return state.getSlashings();
-  }
-
-  @Override
-  public Bitvector getJustification_bits() {
-    return state.getJustification_bits();
-  }
-
-  @Override
-  public Checkpoint getPrevious_justified_checkpoint() {
-    return state.getPrevious_justified_checkpoint();
-  }
-
-  @Override
-  public Checkpoint getCurrent_justified_checkpoint() {
-    return state.getCurrent_justified_checkpoint();
-  }
-
-  @Override
-  public Checkpoint getFinalized_checkpoint() {
-    return state.getFinalized_checkpoint();
+    final int fieldIndex = schema.getFieldIndex(BeaconStateFields.GENESIS_TIME.name());
+    return ((SszPrimitives.SszUInt64) get(fieldIndex)).get();
   }
 
   @Override
   public Optional<SSZList<PendingAttestation>> maybeGetPrevious_epoch_attestations() {
-    return state.maybeGetPrevious_epoch_attestations();
+    return Optional.ofNullable(
+            schema.getField(BeaconStateFields.PREVIOUS_EPOCH_ATTESTATIONS.name()))
+        .map(
+            field ->
+                new SSZBackingList<>(
+                    PendingAttestation.class,
+                    getAny(field.getIndex()),
+                    Function.identity(),
+                    Function.identity()));
+  }
+
+  // TODO implement other methods below
+
+  @Override
+  public Bytes32 getGenesis_validators_root() {
+    return null;
+  }
+
+  @Override
+  public UInt64 getSlot() {
+    return null;
+  }
+
+  @Override
+  public Fork getFork() {
+    return null;
+  }
+
+  @Override
+  public ForkInfo getForkInfo() {
+    return null;
+  }
+
+  @Override
+  public BeaconBlockHeader getLatest_block_header() {
+    return null;
+  }
+
+  @Override
+  public SSZVector<Bytes32> getBlock_roots() {
+    return null;
+  }
+
+  @Override
+  public SSZVector<Bytes32> getState_roots() {
+    return null;
+  }
+
+  @Override
+  public SSZList<Bytes32> getHistorical_roots() {
+    return null;
+  }
+
+  @Override
+  public Eth1Data getEth1_data() {
+    return null;
+  }
+
+  @Override
+  public SSZList<Eth1Data> getEth1_data_votes() {
+    return null;
+  }
+
+  @Override
+  public UInt64 getEth1_deposit_index() {
+    return null;
+  }
+
+  @Override
+  public SSZList<Validator> getValidators() {
+    return null;
+  }
+
+  @Override
+  public SSZList<UInt64> getBalances() {
+    return null;
+  }
+
+  @Override
+  public SSZVector<Bytes32> getRandao_mixes() {
+    return null;
+  }
+
+  @Override
+  public SSZVector<UInt64> getSlashings() {
+    return null;
+  }
+
+  @Override
+  public Bitvector getJustification_bits() {
+    return null;
+  }
+
+  @Override
+  public Checkpoint getPrevious_justified_checkpoint() {
+    return null;
+  }
+
+  @Override
+  public Checkpoint getCurrent_justified_checkpoint() {
+    return null;
+  }
+
+  @Override
+  public Checkpoint getFinalized_checkpoint() {
+    return null;
   }
 
   @Override
   public Optional<SSZList<PendingAttestation>> maybeGetCurrent_epoch_attestations() {
-    return state.maybeGetCurrent_epoch_attestations();
+    return Optional.empty();
   }
 
   @Override
   public Optional<SSZList<SSZVector<SszPrimitives.SszBit>>> maybeGetPreviousEpochParticipation() {
-    return state.maybeGetPreviousEpochParticipation();
+    return Optional.empty();
   }
 
   @Override
   public Optional<SSZList<SSZVector<SszPrimitives.SszBit>>> maybeGetCurrentEpochParticipation() {
-    return state.maybeGetCurrentEpochParticipation();
+    return Optional.empty();
   }
 
   @Override
   public BeaconState update(final Consumer<MutableBeaconState> updater) {
-    return state.update(updater);
+    return null;
   }
 }
