+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/datastructures/state/BeaconStateSchema.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.spec.containers.state;
+package tech.pegasys.teku.spec.datastructures.state;
 
 import java.util.List;
 import java.util.stream.Collectors;
