+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/datastructures/state/genesis/DelegatingBeaconStateGenesis.java
@@ -11,14 +11,14 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.spec.containers.state.genesis;
+package tech.pegasys.teku.spec.datastructures.state.genesis;
 
 import static com.google.common.base.Preconditions.checkArgument;
 
 import java.util.function.Consumer;
 import tech.pegasys.teku.datastructures.state.PendingAttestation;
-import tech.pegasys.teku.spec.containers.state.BeaconState;
-import tech.pegasys.teku.spec.containers.state.DelegatingBeaconState;
+import tech.pegasys.teku.spec.datastructures.state.BeaconState;
+import tech.pegasys.teku.spec.datastructures.state.DelegatingBeaconState;
 import tech.pegasys.teku.ssz.SSZTypes.SSZList;
 
 class DelegatingBeaconStateGenesis extends DelegatingBeaconState implements BeaconStateGenesis {
