+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/schemas/genesis/SchemaDefinitionsGenesis.java
@@ -21,9 +21,9 @@ import tech.pegasys.teku.datastructures.state.Fork;
 import tech.pegasys.teku.datastructures.state.PendingAttestation;
 import tech.pegasys.teku.datastructures.state.Validator;
 import tech.pegasys.teku.spec.constants.SpecConstants;
-import tech.pegasys.teku.spec.containers.state.BeaconState;
-import tech.pegasys.teku.spec.containers.state.BeaconStateFields;
-import tech.pegasys.teku.spec.containers.state.BeaconStateSchema;
+import tech.pegasys.teku.spec.datastructures.state.BeaconState;
+import tech.pegasys.teku.spec.datastructures.state.BeaconStateFields;
+import tech.pegasys.teku.spec.datastructures.state.BeaconStateSchema;
 import tech.pegasys.teku.spec.schemas.SchemaDefinitions;
 import tech.pegasys.teku.ssz.backing.schema.SszComplexSchemas;
 import tech.pegasys.teku.ssz.backing.schema.SszContainerSchema;
